orchestrator/conductor/requirements.txt
langgraph==0.2.74
langchain-core==0.3.51
asyncpg==0.30.0
httpx==0.28.1
python-telegram-bot==21.10
apscheduler==3.11.0
pydantic==2.11.1
pydantic-settings==2.8.1
